home page with @taggelos

+++ src/main/java/gr/uoa/di/repositories/EventRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface EventRepository extends CrudRepository<Event, Long> {
+public interface EventRepository extends CrudRepository<Event, Integer> {
     List<Event> findByCategoryOrTitleOrDescription(String category, String title, String description);
     List<Event> findAll();
     Event findByTitle(String title);
