+++ src/main/java/gr/uoa/di/repositories/ParentMetadataRepository.java
@@ -3,6 +3,6 @@ package gr.uoa.di.repositories;
 import gr.uoa.di.entities.ParentMetadata;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ParentMetadataRepository extends CrudRepository<ParentMetadata, Long> {
+public interface ParentMetadataRepository extends CrudRepository<ParentMetadata, Integer> {
 
 }
