+++ src/main/java/gr/uoa/di/repositories/PlaceRepository.java
@@ -3,6 +3,6 @@ package gr.uoa.di.repositories;
 import gr.uoa.di.entities.Place;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PlaceRepository extends CrudRepository<Place, Long> {
+public interface PlaceRepository extends CrudRepository<Place, Integer> {
 
 }
